refactor: restructure A2A tools and remove deprecated components
- Deleted obsolete files related to A2A SDK invocation and URL validation to streamline the codebase.
- Introduced new modules for building agent tools and handling agent card fetching, enhancing modularity and maintainability.
- Updated index exports to reflect the new structure and ensure proper integration of the A2A SDK functionalities.
- Improved error handling and response parsing in the new implementations for better consistency.
