+++ sites/playground/server/src/a2a-tools/build-tools/index.ts
@@ -1,0 +1,1 @@
+export { buildAgentTools } from './build-agent-tools.js';
