+++ sites/playground/server/src/a2a-tools/fetch-card/handler.ts
@@ -1,18 +1,10 @@
 import type { Request, Response as ExpressResponse } from 'express';
 import getRawBody from 'raw-body';
-import { isAllowedAgentUrl, isPlaygroundDevelopment } from './agent-url-validation.js';
-import { normalizeAgentCard } from './protocol/supported-interfaces.js';
+import { normalizeAgentCard } from '../parse-card/parse.js';
+import { isAllowedAgentUrl, isPlaygroundDevelopment } from '../guard-agent-url/guard.js';
 
 /** 拉取远程 Agent Card 的单次请求超时（毫秒）。 */
 const UPSTREAM_FETCH_TIMEOUT_MS = 10_000;
-
-function sendAgentCardResponse(
-  res: ExpressResponse,
-  httpStatus: number,
-  body: { data?: unknown; message?: string },
-): void {
-  res.status(httpStatus).json(body);
-}
 
 /**
  * 服务端代理拉取 Agent Card，规避浏览器跨域限制，并规范化 api.url 字段。
@@ -27,19 +19,19 @@ export const fetchAgentCardHandler = async (req: Request, res: ExpressResponse):
     const requestedUrl = (body?.url || '').trim();
 
     if (!requestedUrl) {
-      sendAgentCardResponse(res, 400, { message: '缺少 Agent Card URL' });
+      res.status(400).json({ message: '缺少 Agent Card URL' });
       return;
     }
 
     try {
       new URL(requestedUrl);
     } catch {
-      sendAgentCardResponse(res, 400, { message: 'Agent Card URL 无效' });
+      res.status(400).json({ message: 'Agent Card URL 无效' });
       return;
     }
 
     if (!isPlaygroundDevelopment && !isAllowedAgentUrl(requestedUrl)) {
-      sendAgentCardResponse(res, 403, { message: '不允许访问本地或内网地址' });
+      res.status(403).json({ message: '不允许访问本地或内网地址' });
       return;
     }
 
@@ -55,7 +47,7 @@ export const fetchAgentCardHandler = async (req: Request, res: ExpressResponse):
       });
     } catch (error: any) {
       const message = error?.name === 'AbortError' ? '获取 Agent Card 超时' : error?.message || String(error);
-      sendAgentCardResponse(res, 502, { message });
+      res.status(502).json({ message });
       return;
     } finally {
       clearTimeout(timeoutId);
@@ -64,7 +56,7 @@ export const fetchAgentCardHandler = async (req: Request, res: ExpressResponse):
     const rawText = await fetchRes.text();
 
     if (!fetchRes.ok) {
-      sendAgentCardResponse(res, fetchRes.status >= 400 ? fetchRes.status : 502, {
+      res.status(fetchRes.status >= 400 ? fetchRes.status : 502).json({
         message: rawText.trim() || `HTTP ${fetchRes.status} ${fetchRes.statusText}`.trim(),
       });
       return;
@@ -74,20 +66,20 @@ export const fetchAgentCardHandler = async (req: Request, res: ExpressResponse):
     try {
       card = rawText.trim() ? JSON.parse(rawText) : null;
     } catch {
-      sendAgentCardResponse(res, 500, { message: 'Agent Card 响应不是有效 JSON' });
+      res.status(500).json({ message: 'Agent Card 响应不是有效 JSON' });
       return;
     }
 
     if (!card || typeof card !== 'object' || Array.isArray(card)) {
-      sendAgentCardResponse(res, 500, { message: 'Agent Card 格式无效' });
+      res.status(500).json({ message: 'Agent Card 格式无效' });
       return;
     }
 
     const normalizedCard = normalizeAgentCard(card as Record<string, unknown>);
-    sendAgentCardResponse(res, 200, { data: normalizedCard });
+    res.status(200).json({ data: normalizedCard });
   } catch (error: any) {
     const message = error?.message || String(error);
     const httpStatus = error?.name === 'AgentCardProtocolError' ? 422 : 500;
-    sendAgentCardResponse(res, httpStatus, { message });
+    res.status(httpStatus).json({ message });
   }
 };
