+++ sites/playground/server/src/a2a-tools/fetch-card/index.ts
@@ -1,0 +1,1 @@
+export { fetchAgentCardHandler } from './handler.js';
