+++ sites/playground/server/src/a2a-tools/invoke-agent/index.ts
@@ -1,0 +1,1 @@
+export { invokeA2aAgent } from './invoke-a2a-agent.js';
