--- sites/playground/server/src/a2a-tools/protocol/adapters/index.ts
@@ -1,2 +1,0 @@
-export { toAgentCard } from './agent-card.js';
-export type { A2aInvokeAdapterContext, A2aProtocolInvokeAdapter } from './types.js';
