--- sites/playground/server/src/a2a-tools/protocol/index.ts
@@ -1,12 +1,0 @@
-export { toAgentCard } from './adapters/index.js';
-export type { A2aInvokeAdapterContext, A2aProtocolInvokeAdapter } from './adapters/index.js';
-export {
-  AgentCardProtocolError,
-  normalizeAgentCard,
-  resolveAgentApiUrl,
-  resolveAgentInterface,
-  tryResolveAgentInterface,
-  type ResolvedAgentInterface,
-} from './supported-interfaces.js';
-export { A2A_PROTOCOL_CONFIG } from './types.js';
-export type { A2aProtocolBinding, A2aProtocolVersion, AgentProtocolSource } from './types.js';
